Working on the interface
At first, elements were on the top left corner, and out of the canvas because they were coded as HTML elements. I am integrating them directly in the canvas with a p5.js code instead. I placed the timer and water meter and moved the waterfall to the right so I can put the text box to the left.