It is now impossible to generate a key outside of the box
Before, if you just pressed any key, the water would fall, even if you're within the box or not.I will try a second version, but this is the first version which prevents that. The only problem is that your mouse has to be inside the box at all times.